+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - Rendering backend architecture to a more flexible one.
 - Plugin API to allow calls from multiple threads.
 - Made `widget-control property-update` update the ewwii property instea of the gtk property.
+- `ipc_request` plugin API to return `FutureResult`.
 
 ### Removed